app/requirements.txt
asgiref==3.5.0
Django==4.0
django-filter==21.1
django-leaflet==0.28.2
djangorestframework==3.13.1
djangorestframework-gis==0.18
psycopg2-binary==2.8.6
pytz==2021.3
sqlparse==0.4.2
